+++ app/src/ui/calendar_widget.h
@@ -3,6 +3,7 @@
 
 #include <QDialog>
 #include <QDate>
+#include "../constants.h"
 
 namespace Ui {
 class CalendarWidget;
@@ -20,7 +21,7 @@ class CalendarWidget : public QDialog
     Q_OBJECT
 
 public:
-    explicit CalendarWidget(int row, int varenr, const QString& lokasjon, QWidget *parent = nullptr);
+    explicit CalendarWidget(int row, int varenr, const QString& lokasjon, const QString& saved_date, QWidget *parent = nullptr);
     ~CalendarWidget();
 
 signals:
@@ -33,6 +34,7 @@ private:
     int m_row;
     int m_varenr;
     QString m_lokasjon;
+    QString m_saved_date;
 };
 
 } // namespace