+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.4.0
+- Removed the save button, its not used anymore. You can now double click to update the location and date
+- Date is passed to the widget when you updating the location or date.
+- Whole line is now in color.
+
 ## v0.3.0
 - You can now scan 2d matrix codes (FMD) and it will automaticly add the product to the durability list with durability date.
 - Double clicking the products will open a new window where you can change location and date.